chore(docs): adapt docs layout, rename stories files, add documentation
