+++ .storybook/About.mdx
@@ -1,0 +1,16 @@
+# EOxElements [beta]
+
+[GitHub repository](https://github.com/EOX-A/EOxElements)
+
+## Introduction
+EOxElements are a set of reusable, modular and highly configurable components for building geospatial applications.
+The goal is to reduce code duplication during development, and offer ready-made solutions for quick bootstrapping and
+increased long-term stabilty.
+
+## Technology
+The EOxElements are a suite of custom elements (a.k.a. [web components](https://www.webcomponents.org/introduction)), which are framework-agnostic.
+Thus, they can be used in applications built with Vanilla JavaScript, Vue.js, React, Angular etc. 
+
+## Current state
+All of the elements are currently functional, but not yet ready for production-use because of frequent breaking changes.
+Please refer to the repository's [Milestones](https://github.com/EOX-A/EOxElements/milestones) to get a feeling of when to expect a producion-ready v1.0.0.
