+++ .storybook/DocumentationTemplate.mdx
@@ -1,0 +1,26 @@
+{/* DocumentationTemplate.mdx */}
+
+import {
+  Meta,
+  Title,
+  Subtitle,
+  Description,
+  Primary,
+  Controls,
+  Stories,
+} from "@storybook/blocks";
+
+<Meta isTemplate />
+
+<Title />
+<Subtitle />
+
+<Primary />
+
+<Description />
+
+### API
+
+<Controls />
+
+<Stories />
