+++ .storybook/main.js
@@ -1,6 +1,9 @@
 /** @type { import('@storybook/web-components-vite').StorybookConfig } */
 const config = {
-  stories: ["../**/stories.@(js|jsx|mjs|ts|tsx)"],
+  stories: [
+    "./About.mdx",
+    "../**/*.stories.@(js|jsx|mjs|ts|tsx)",
+  ],
   addons: ["@storybook/addon-essentials"],
   framework: "@storybook/web-components-vite",
   docs: {
