+++ .storybook/preview-head.html
@@ -1,0 +1,6 @@
+<style>
+  @import url("https://eox.at/fonts/roboto/roboto.css");
+  * {
+    font-family: "Roboto", sans-serif;
+  }
+</style>
